+++ bfabric/src/bfabric/operations/dataset/_column_types.py
@@ -1,0 +1,21 @@
+from __future__ import annotations
+
+import importlib.resources
+
+import yaml
+from pydantic import BaseModel
+
+
+class DatasetColumnTypes(BaseModel):
+    entities: set[str]
+    """List of dataset column types for referencing entities, e.g. Resource, Dataset, Sample."""
+
+
+class DatasetColumnTypesFile(BaseModel):
+    dataset_column_types: DatasetColumnTypes
+
+
+def get_dataset_column_types() -> DatasetColumnTypes:
+    """Parses the default `_column_types.yml` file from the package."""
+    with importlib.resources.open_text("bfabric.operations.dataset", "_column_types.yml") as f:
+        return DatasetColumnTypesFile.model_validate(yaml.safe_load(f)).dataset_column_types